Improve reusability of the /uninstall route - Pass the entire user object to MailiumUninstallEvent - Delete the user after the event is handled This allows app developers to still read the data related to the user, and for instance use it to deauthorize the user on the connected platform.

+++ http/routes.php
@@ -15,11 +15,13 @@ Route::post('/uninstall', ['middleware' => [
     /** @var MailiumAppAuthenticatable $user */
     $user = $request->attributes->get('mailium_app_user');
 
-    // Delete user from database;
-    $user->delete();
-
     // Fire uninstall event
-    Event::fire(new MailiumAppUninstallEvent($user->accid));
+    Event::fire(new MailiumAppUninstallEvent($user));
+
+    // Delete user from database;
+    if ($user->exists) {
+        $user->delete();
+    }
 
     return response()->json([], 200);
 }])->name('mailium_app_uninstall');
